move getLessonType() to vue
